major changes to composition structure
