+++ docs/composition.rst
@@ -3,7 +3,18 @@
 xdust.graindist.composition
 ===========================
 
-Superclass: ``Composition``
+The composition library is built on the superclass :ref:`CompositionClass`, which holds the optical constants and material density [g cm^-3] of a grain composition type.
+
+**Premade composition types**
+
+* :ref:`CmDrude` -- Drude approximation, treating the grain as a mass of free electrons
+* :ref:`CmSilicate` -- astronomical silicates (Draine 2003)
+* :ref:`CmGraphite` -- graphite (Draine 2003)
+
+
+.. _CompositionClass:
+
+Composition
 ---------------------------
 
 .. autoclass:: xdust.graindist.composition.Composition
