+++ src/xdust/graindist/composition/cmdrude.py
@@ -2,7 +2,7 @@ import numpy as np
 import astropy.units as u
 import astropy.constants as c
 
-from .composition import Composition
+from . import Composition
 
 __all__ = ['CmDrude']
